refactor: split engine.rs into three files
