+++ src/flow/src/batching_mode.rs
@@ -18,9 +18,21 @@ use std::time::Duration;
 
 mod engine;
 mod frontend_client;
+mod state;
+mod task;
 mod time_window;
 mod utils;
 
 /// TODO(discord9): make those constants configurable
 /// The default batching engine query timeout is 10 minutes
 pub const DEFAULT_BATCHING_ENGINE_QUERY_TIMEOUT: Duration = Duration::from_secs(10 * 60);
+
+/// TODO(discord9): make those constants configurable
+/// The default rule engine query timeout is 10 minutes
+pub const DEFAULT_RULE_ENGINE_QUERY_TIMEOUT: Duration = Duration::from_secs(10 * 60);
+
+/// will output a warn log for any query that runs for more that 1 minutes, and also every 1 minutes when that query is still running
+pub const SLOW_QUERY_THRESHOLD: Duration = Duration::from_secs(60);
+
+/// The minimum duration between two queries execution by batching mode task
+const MIN_REFRESH_DURATION: Duration = Duration::new(5, 0);
